1064: WIP - added notes for some potential refactor work [skip ci]
